Move cataloglayerselected event handler controls from mapservice.js to map-controls.js store
